Backend/requirements.txt
flask
flask-cors
flask_sqlalchemy
bcrypt
python-dotenv
mysql-connector-python
